+++ packages/core/src/lib/cli/internal/reporter/supported-reporter-formats.ts
